babysitter/requirements.txt
PyYAML>=6.0.1
pydantic>=2.12.5